chore: make the `channelElements` private (#11)
## Description

This PR removes a public `channelElement` from the public `RSS`
interface. Doing so allows the implementation of the `RSS` interface
(the `Feed` class in this context) to make its `channelElements` private
and inaccessible to the public.

The minor improvement should make usage of the code much simpler and
easier to use.

## Type of Change

Select the type of change this PR introduces:

- [ ] Feature
- [ ] Bug Fix
- [x] Refactor
- [ ] Documentation Update
- [ ] Other (please describe below)

## How Has This Been Tested?

Describe the testing strategy and steps taken to verify your changes.