ci: automate the version releases (#13)
## Description

This PR adds a new GitHub Action workflow to automate the process of
making a stable version release of the software. The workflow will:

1. Run all sorts of QA checks (including ESLint and Prettier).
2. Run the test suite using Vitest.
3. Make a release on GitHub (along with the generated release notes).
4. Publish the latest version on the NPM package registry.

The first set of two jobs (the ESLint and Prettier jobs) and the second
set of next two jobs (GitHub release and NPM package registry job) run
in parallel but the second set of jobs are dependent on the QA jobs.

**NOTE**: In addition to the new updates, the previous workflow for
automating the pre-releases were also refactored slightly for some bug
fixes and improvements in legibility.

## Type of Change

- [ ] Feature
- [ ] Bug Fix
- [ ] Refactor
- [ ] Documentation Update
- [x] Other (please describe below)