fix: render CDATA instead of plain text (#22)
The current implementation of the generator provides a string of XML
data with fields in plain text. This data can be misinterpreted and
rendered improperly by RSS clients. The solution to the issue is to
render the XML data using CDATA. See this relevant thread on Stack
Overflow - [What does `<![CDATA[]]>` in XML
mean?](https://stackoverflow.com/q/2784183/8604951) to learn more about
the importance of using CDATA for the rendered XML data.

The current implementation produces the following output:

```xml
<?xml version="1.0"?>
<rss version="2.0">
  <channel>
    <title>Lorem Ipsum</title>
    <link>https://jarmos.dev/rss.xml</link>
    <description>Lorem Ipsum</description>
    <item>
      <title>Lorem Ipsum</title>
      <description>Lorem Ipsum</description>
    </item>
  </channel>
</rss>
```

The new implementation produces the following proper output:

```xml
<?xml version="1.0"?>
<rss version="2.0">
  <channel>
    <title><![CDATA[Lorem Ipsum]]></title>
    <link><![CDATA[https://jarmos.dev/rss.xml]]></link>
    <description><![CDATA[Lorem Ipsum]]></description>
    <item>
      <title><![CDATA[Lorem Ipsum]]></title>
      <description><![CDATA[Lorem Ipsum]]></description>
    </item>
  </channel>
</rss>
```

Select the type of change this PR introduces:

- [ ] Feature
- [x] Bug Fix
- [ ] Refactor
- [ ] Documentation Update
- [ ] Other (please describe below)