ci: refactor the events triggering the CI pipelines (#30)
## Description

This PR provides some fixes and refactors to trigger the CI **ONLY** in
the event of creating a new PR with changes to the `.ts` (TypeScript)
files. This should avoid triggering the CI unnecessarily for some
arbitrary PRs (e.g. PRs to update the docs).

Including the changes mentioned above, the PR also fixes the
configurations wherein Dependabot PRs were also triggering the CI.

## Type of Change

Select the type of change this PR introduces:

- [ ] Feature
- [x] Bug Fix
- [ ] Refactor
- [ ] Documentation Update
- [ ] Other (please describe below)